+++ src/Symfony/Component/Config/Definition/Builder/TreeBuilder.php
@@ -24,36 +24,10 @@ class TreeBuilder implements NodeParentInterface
     protected $tree;
     protected $root;
 
-    public function __construct(string $name = null, string $type = 'array', NodeBuilder $builder = null)
+    public function __construct(string $name, string $type = 'array', NodeBuilder $builder = null)
     {
-        if (null === $name) {
-            @trigger_error('A tree builder without a root node is deprecated since Symfony 4.2 and will not be supported anymore in 5.0.', E_USER_DEPRECATED);
-        } else {
-            $builder = $builder ?: new NodeBuilder();
-            $this->root = $builder->node($name, $type)->setParent($this);
-        }
-    }
-
-    /**
-     * Creates the root node.
-     *
-     * @param string      $name    The name of the root node
-     * @param string      $type    The type of the root node
-     * @param NodeBuilder $builder A custom node builder instance
-     *
-     * @return ArrayNodeDefinition|NodeDefinition The root node (as an ArrayNodeDefinition when the type is 'array')
-     *
-     * @throws \RuntimeException When the node type is not supported
-     *
-     * @deprecated since Symfony 4.3, pass the root name to the constructor instead
-     */
-    public function root($name, $type = 'array', NodeBuilder $builder = null)
-    {
-        @trigger_error(sprintf('The "%s()" method called for the "%s" configuration is deprecated since Symfony 4.3, pass the root name to the constructor instead.', __METHOD__, $name), E_USER_DEPRECATED);
-
         $builder = $builder ?: new NodeBuilder();
-
-        return $this->root = $builder->node($name, $type)->setParent($this);
+        $this->root = $builder->node($name, $type)->setParent($this);
     }
 
     /**